backend/ebus_backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
breadability==0.1.20
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contourpy==1.2.0
contractions==0.1.73
cryptography==41.0.3
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.3
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-allauth==0.55.2
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-filter==24.2
django-otp==1.5.0
django-session-timeout==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
drf-yasg==1.21.8
eli5==0.11.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
environ==1.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
floret==0.10.5
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
funcy==2.0
gast==0.6.0
gensim==4.3.2
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.62.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.7
imageio==2.34.0
imbalanced-learn==0.12.2
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.7
MDSdata==0.1.4
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
murmurhash==1.0.10
mysqlclient==2.2.3
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyahocorasick==2.1.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycountry==23.12.11
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.8.0
pyLDAvis==3.4.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyphen==0.14.0
pyspellchecker==0.8.1
pysrt==1.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==2.5.1
setuptools==69.2.0
shap==0.45.0
six==1.16.0
slicer==0.0.7
smart-open==6.4.0
sniffio==1.3.1
snorkel==0.10.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
stripe==11.4.1
summa==1.2.0
sumy==0.11.0
sweetify==2.3.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
terminado==0.18.1
textacy==0.13.0
textblob==0.18.0.post0
textsearch==0.0.24
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.38.2
twilio==9.0.5
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
vboxapi==1.0
virtualenv==20.26.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xgboost==2.1.1
xxhash==3.4.1
yarl==1.9.4
yellowbrick==1.5
